Rename `src/ext.rs` to `src/transport_ext.rs`.
